change emdedding type from long to float to vaoid getting al zeros embedding (#2292)

+++ speechbrain/lobes/models/huggingface_transformers/discrete_wav2vec2.py
@@ -161,7 +161,7 @@ class DiscreteWav2Vec2(Wav2Vec2):
         return (
             torch.tensor(
                 embs.reshape(wav.shape[0], -1, embs.shape[-1]),
-                dtype=torch.long,
+                dtype=torch.float,
                 device=wav.device,
             ),
             torch.tensor(
